Moved to gulp task
